AttachedFileRepository integer -> Long

+++ src/main/java/com/example/appcenter_project/repository/announcement/AttachedFileRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface AttachedFileRepository extends JpaRepository<AttachedFile,Integer> {
+public interface AttachedFileRepository extends JpaRepository<AttachedFile, Long> {
     List<AttachedFile> findByAnnouncement(Announcement announcement);
     void deleteByFilePath(String filePath);
